pyproject.toml
[project]
name = "sample-datafold-dbt-project"

version = "0.0.1"

description = "sample-datafold-dbt-project"

readme = "README.md"

requires-python = "3.11"

classifiers = [
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3 :: Only",
]

dependencies = [
    "pandas>=1,<3",
    "requests>=2,<3",
    "dbt-snowflake==1.6.2",
    "sqlfluff==2.3.1",
    "sqlfluff-templater-dbt==2.3.1",
]


[project.optional-dependencies]
dev = ["check-manifest"]
test = ["pytest", "pytest-mock"]


[project.urls]
"Homepage" = "https://github.com/Big-Little-Robots-Org/sample-datafold-dbt-project"
"Source" = "https://github.com/Big-Little-Robots-Org/sample-datafold-dbt-project"

[build-system]
requires = ["setuptools>=43.0.0", "wheel"]
build-backend = "setuptools.build_meta"
